pyproject.toml
[project]
name = "helix-voronoi-workspace"
version = "0.1.0"
description = "Helix Voronoi geometry and topology-backfill workflows"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy-stl>=3.2.0",
    "scipy>=1.17.1",
    "sfepy>=2025.4",
]

[project.scripts]
helix-voronoi = "helix_voronoi.cli:main"
topopt-backfill = "topopt_backfill.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/helix_voronoi", "src/topopt_backfill"]
